apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]
SQLAlchemy==2.0.34
psycopg[binary]==3.2.10
geoalchemy2==0.15.2
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.7
python-socketio[asgi]==5.11.4
minio==7.2.7
python-multipart==0.0.9